email integrated into create reservation process
