+++ app/controllers/index.server.controller.js
@@ -1,13 +1,3 @@
-/*const nodemail = require('../libs/mail');
-const config = require('../../config/config');
-
-var mailOptions = {
-  from: config.email,
-  to: config.email,
-  subject: 'Sending Email using Node.js',
-  text: 'That was easy!'
-};*/
-
 // Create a new 'render' controller method
 exports.render = function (req, res) {
   // Use the 'response' object to render the 'index' view with a 'title' and 'userFullName' properties
@@ -16,6 +6,5 @@ exports.render = function (req, res) {
     headline: 'Reservation System',
     userFullName: req.user ? req.user.fullName : '',
     messages: req.flash('error').concat(req.flash('info'))
-    //nodemail.sendMail(mailOptions)
   });
 };
