+++ package.json
@@ -14,6 +14,7 @@
     "express": "^4.16.3",
     "express-session": "^1.15.6",
     "googleapis": "^27.0.0",
+    "nodemailer": "^4.6.4",
     "method-override": "^2.3.10",
     "mkdirp": "0.5.1",
     "mongoose": "^5.0.14",
